fixed a rename issue from long past
