+++ spiNNakerGraphFrontEnd.cfg
@@ -81,6 +81,7 @@ turn_off_machine = True
 clear_routing_tables = True
 clear_tags = True
 enable_reinjection = True
+machine_spec_file = None
 ResetMachineOnStartupFlag = True
 
 [Mapping]
