+++ spinnaker_graph_front_end/abstract_partitioned_data_specable_vertex.py
