+++ spinnaker_graph_front_end/extra_pacman_algorithms/spinnaker_graph_front_end_edge_to_key_mapper.py
