+++ spinnaker_graph_front_end/extra_pacman_algorithms/spinnaker_graph_front_end_partitioned_graph_host_based_data_specification_executor.py
@@ -5,7 +5,7 @@ from spinn_front_end_common.abstract_models.\
     AbstractDataSpecableVertex
 from spinn_front_end_common.utilities import exceptions as front_end_exceptions
 
-from spynnaker_graph_front_end.\
+from spinnaker_graph_front_end.\
     abstract_partitioned_data_specable_vertex import \
     AbstractPartitionedDataSpecableVertex
